+++ src/root/longdouble.h
@@ -18,7 +18,7 @@
 typedef real_t longdouble;
 
 template<typename T> longdouble ldouble(T x) { return (longdouble) x; }
-inline int ld_sprint(char* str, int fmt, longdouble x)
+inline size_t ld_sprint(char* str, int fmt, longdouble x)
 {
     if(fmt == 'a' || fmt == 'A')
         return x.formatHex(buffer, 46); // don't know the size here, but 46 is the max
@@ -34,7 +34,7 @@ typedef volatile long double volatile_longdouble;
 // template<typename T> longdouble ldouble(T x) { return (longdouble) x; }
 #define ldouble(x) ((longdouble)(x))
 
-inline int ld_sprint(char* str, int fmt, longdouble x)
+inline size_t ld_sprint(char* str, int fmt, longdouble x)
 {
     char sfmt[4] = "%Lg";
     sfmt[2] = fmt;
@@ -247,7 +247,7 @@ public:
 //_STCONSDEF(numeric_limits<longdouble>, int, min_exponent)
 //_STCONSDEF(numeric_limits<longdouble>, int, min_exponent10)
 
-int ld_sprint(char* str, int fmt, longdouble x);
+size_t ld_sprint(char* str, int fmt, longdouble x);
 
 #endif // !_MSC_VER
